+++ backend/database.py
@@ -1,0 +1,43 @@
+import sentry_sdk
+from apscheduler.schedulers.background import BackgroundScheduler
+from fastapi.middleware.cors import CORSMiddleware
+from sqlalchemy.orm import sessionmaker
+from fastapi import FastAPI
+from sqlalchemy import (create_engine)
+from sqlalchemy.ext.declarative import declarative_base
+from pydantic import BaseModel
+
+Base = declarative_base()
+engine = create_engine("sqlite:///news_database.db", echo=True)
+
+Base.metadata.create_all(engine)
+
+Session = sessionmaker(bind=engine)
+
+sentry_sdk.init(
+    dsn="https://4001ffe917ccb261aa0e0c34026dc343@o4505702629834752.ingest.us.sentry.io/4507694792704000",
+    traces_sample_rate=1.0,
+    profiles_sample_rate=1.0,
+)
+
+app = FastAPI()
+bgs = BackgroundScheduler()
+SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
+
+app.add_middleware(
+    CORSMiddleware,  # noqa
+    allow_origins=["http://localhost:8080"],
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+class UserAuthSchema(BaseModel):
+    username: str
+    password: str
+
+class NewsSumaryRequestSchema(BaseModel):
+    content: str
+    
+class PromptRequest(BaseModel):
+    prompt: str
