+++ backend/get/news.py
@@ -1,0 +1,39 @@
+from ..Loginverification import Loginverification
+from ..models import NewsArticle,user_news_association_table
+from ..database import app
+
+from fastapi import APIRouter,Depends
+
+routes=APIRouter()
+@app.get("/api/v1/news/news")
+def read_news(db=Depends(Loginverification.session_opener)):
+    """
+    read new
+
+    :param db:
+    :return:
+    """
+    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
+    result = []
+    for n in news:
+        upvotes, upvoted = get_article_upvote_details(n.id, None, db)
+        result.append(
+            {**n.__dict__, "upvotes": upvotes, "is_upvoted": upvoted}
+        )
+    return result
+
+def get_article_upvote_details(article_id, uid, db):
+    cnt = (
+        db.query(user_news_association_table)
+        .filter_by(news_articles_id=article_id)
+        .count()
+    )
+    voted = False
+    if uid:
+        voted = (
+                db.query(user_news_association_table)
+                .filter_by(news_articles_id=article_id, user_id=uid)
+                .first()
+                is not None
+        )
+    return cnt, voted
