+++ backend/get/price.py
@@ -1,0 +1,14 @@
+from ..database import app
+
+import requests
+from fastapi import  APIRouter,Query
+
+routes=APIRouter()
+@app.get("/api/v1/prices/necessities-price")
+def get_necessities_prices(
+        category=Query(None), commodity=Query(None)
+):
+    return requests.get(
+        "https://opendata.ey.gov.tw/api/ConsumerProtection/NecessitiesPrice",
+        params={"CategoryName": category, "Name": commodity},
+    ).json()
