+++ backend/get/usernews.py
@@ -1,0 +1,47 @@
+from ..database import app
+from ..Loginverification import Loginverification
+from ..models import NewsArticle,user_news_association_table
+
+from fastapi import APIRouter,Depends
+
+routes=APIRouter()
+@app.get("/api/v1/news/user_news")
+def read_user_news(
+        db=Depends(Loginverification.session_opener),
+        user=Depends(Loginverification.authenticate_user_token)
+):
+    """
+    read user new
+
+    :param db:
+    :param u:
+    :return:
+    """
+    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
+    result = []
+    for article in news:
+        upvotes, upvoted = get_article_upvote_details(article.id, user.id, db)
+        result.append(
+            {
+                **article.__dict__,
+                "upvotes": upvotes,
+                "is_upvoted": upvoted,
+            }
+        )
+    return result
+
+def get_article_upvote_details(article_id, uid, db):
+    cnt = (
+        db.query(user_news_association_table)
+        .filter_by(news_articles_id=article_id)
+        .count()
+    )
+    voted = False
+    if uid:
+        voted = (
+                db.query(user_news_association_table)
+                .filter_by(news_articles_id=article_id, user_id=uid)
+                .first()
+                is not None
+        )
+    return cnt, voted
