+++ backend/main.py
@@ -1,563 +1,51 @@
-import json
-import sentry_sdk
+from fastapi import FastAPI
 from apscheduler.schedulers.background import BackgroundScheduler
-from fastapi.middleware.cors import CORSMiddleware
-import itertools
-from sqlalchemy import delete, insert, select
-from sqlalchemy.orm import Session, sessionmaker
-from typing import List, Optional
-import requests
-from fastapi import APIRouter, HTTPException, Query, Depends, status, FastAPI
-import os
-from datetime import datetime, timedelta
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jose import JWTError, jwt
-from passlib.context import CryptContext
 
-from pydantic import BaseModel, Field, AnyHttpUrl
-from sqlalchemy import (Column, ForeignKey, Integer, String, Table, Text,
-                        create_engine)
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import relationship, sessionmaker
+from get.me import routes as me_routes
+from get.price import routes as price_routes
+from get.news import routes as news_routes
+from get.usernews import routes as usernews_routes
+from post.login import routes as login_routes
+from post.summary import routes as summary_routes
+from post.search import routes as search_routes
+from post.register import routes as register_routes
+from post.upvote import routes as upvote_routes
+from database import Base, engine, bgs, SessionLocal
+from NewsService import NewsService
+from models import NewsArticle
 
-Base = declarative_base()
+# 建立資料表
+Base.metadata.create_all(bind=engine)
 
-
-user_news_association_table = Table(
-    "user_news_upvotes",
-    Base.metadata,
-    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
-    Column(
-        "news_articles_id", Integer, ForeignKey("news_articles.id"), primary_key=True
-    ),
-)
-
-# from pydantic import BaseModel
-
-
-class User(Base):
-    __tablename__ = "users"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    username = Column(String(50), unique=True, nullable=False)
-    hashed_password = Column(String(200), nullable=False)
-    upvoted_news = relationship(
-        "NewsArticle",
-        secondary=user_news_association_table,
-        back_populates="upvoted_by_users",
-    )
-
-
-class NewsArticle(Base):
-    __tablename__ = "news_articles"
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    url = Column(String, unique=True, nullable=False)
-    title = Column(String, nullable=False)
-    time = Column(String, nullable=False)
-    content = Column(Text, nullable=False)
-    summary = Column(Text, nullable=False)
-    reason = Column(Text, nullable=False)
-    upvoted_by_users = relationship(
-        "User", secondary=user_news_association_table, back_populates="upvoted_news"
-    )
-
-
-engine = create_engine("sqlite:///news_database.db", echo=True)
-
-Base.metadata.create_all(engine)
-
-Session = sessionmaker(bind=engine)
-
-sentry_sdk.init(
-    dsn="https://4001ffe917ccb261aa0e0c34026dc343@o4505702629834752.ingest.us.sentry.io/4507694792704000",
-    traces_sample_rate=1.0,
-    profiles_sample_rate=1.0,
-)
-
+# 建立 FastAPI app
 app = FastAPI()
-bgs = BackgroundScheduler()
-SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-app.add_middleware(
-    CORSMiddleware,  # noqa
-    allow_origins=["http://localhost:8080"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-import os
-from openai import OpenAI
-
-
-# def generate_summary(content):
-#     m = [
-#         {
-#             "role": "system",
-#             "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-#         },
-#         {"role": "user", "content": f"{content}"},
-#     ]
-#
-#     completion = OpenAI(api_key="xxx").chat.completions.create(
-#         model="gpt-3.5-turbo",
-#         messages=m,
-#     )
-#     return completion.choices[0].message.content
-
-#
-# def extract_search_keywords(content):
-#     m = [
-#         {
-#             "role": "system",
-#             "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
-#         },
-#         {"role": "user", "content": f"{content}"},
-#     ]
-#
-#     completion = OpenAI(api_key="xxx").chat.completions.create(
-#         model="gpt-3.5-turbo",
-#         messages=m,
-#     )
-#     return completion.choices[0].message.content
-
-
-from urllib.parse import quote
-import requests
-from bs4 import BeautifulSoup
-from sqlalchemy.orm import Session
-
-class NewsService:
-    def __init__(self, db_session: Session, ai_api_key: str):
-        """
-        db_session: SQLAlchemy Session
-        ai_api_key: OpenAI API Key
-        """
-        self.db = db_session
-        self.ai_key = ai_api_key
-        self.ai_client = OpenAI(api_key=self.ai_key)
-
-    def add_new(self,news_data):
-        """
-        add new to db
-        :param news_data: news info
-        :return:
-        """
-        self.db.add(NewsArticle(
-            url=news_data["url"],
-            title=news_data["title"],
-            time=news_data["time"],
-            content=" ".join(news_data["content"]),  # 將內容list轉換為字串
-            summary=news_data["summary"],
-            reason=news_data["reason"],
-        ))
-        self.db.commit()
-        self.db.close()
-
-
-    def get_new_info(self,search_term, is_initial=False):
-        """
-        get new
-
-        :param search_term:
-        :param is_initial:
-        :return:
-        """
-        all_news_data = []
-        # iterate pages to get more news data, not actually get all news data
-        if is_initial:
-            for page in range(1, 10):
-                params = {
-                    "page": page,
-                    "id": f"search:{quote(search_term)}",
-                    "channelId": 2,
-                    "type": "searchword",
-                }
-                response = requests.get("https://udn.com/api/more", params=params)
-
-                all_news_data.append(response.json()["lists"])
-                
-        else:
-            params = {
-                "page": 1,
-                "id": f"search:{quote(search_term)}",
-                "channelId": 2,
-                "type": "searchword",
-            }
-            response = requests.get("https://udn.com/api/more", params=params)
-
-            all_news_data = response.json()["lists"]
-
-        return all_news_data
-
-    def evaluate_relevance(self,title: str) -> str:
-        """使用 OpenAI 評估新聞標題是否與主題有關"""
-        messages = [
-            {
-                "role": "system",
-                "content": (
-                    "你是一個關聯度評估機器人，請評估新聞標題是否與「民生用品的價格變化」相關，"
-                    "並給予'high'、'medium'、'low'評價。(僅需回答'high'、'medium'、'low'三個詞之一)"
-                ),
-            },
-            {"role": "user", "content": title},
-        ]
-
-        ai = OpenAI(api_key="xxx").chat.completions.create(
-            model="gpt-3.5-turbo", 
-            messages=messages,
-        )
-        return ai.choices[0].message.content.strip().lower()
-    def process_news_article(self,news):
-        response = requests.get(news["titleLink"])
-        soup = BeautifulSoup(response.text, "html.parser")
-        # 標題
-        title = soup.find("h1", class_="article-content__title").text
-        time = soup.find("time", class_="article-content__time").text
-        # 定位到包含文章内容的 <section>
-        content_section = soup.find("section", class_="article-content__editor")
-
-        paragraphs = [
-            p.text
-            for p in content_section.find_all("p")
-            if p.text.strip() != "" and "▪" not in p.text
-        ]
-        detailed_news =  {
-            "url": news["titleLink"],
-            "title": title,
-            "time": time,
-            "content": paragraphs,
-        }
-        messages = [
-            {
-                "role": "system",
-                "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-            },
-            {
-                "role": "user", 
-                "content": " ".join(detailed_news["content"])
-            },
-        ]
-
-        completion = OpenAI(api_key="xxx").chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=messages,
-        )
-        result = completion.choices[0].message.content
-        result = json.loads(result)
-        detailed_news["summary"] = result["影響"]
-        detailed_news["reason"] = result["原因"]
-        self.add_new(detailed_news)
-    def get_new(self,is_initial=False):
-        """
-        get new info
-
-        :param is_initial:
-        :return:
-        """
-        news_data = self.get_new_info("價格", is_initial=is_initial)
-        for news in news_data:
-            relevance =self.evaluate_relevance(news["title"])
-            if relevance == "high":
-                self.process_news_article(news)
-
+# include 各路由
+app.include_router(me_routes)
+app.include_router(price_routes)
+app.include_router(news_routes)
+app.include_router(usernews_routes)
+app.include_router(login_routes)
+app.include_router(summary_routes)
+app.include_router(search_routes)
+app.include_router(register_routes)
+app.include_router(upvote_routes)
 
 
 @app.on_event("startup")
 def start_scheduler():
     db = SessionLocal()
-    news_service = NewsService(db_session=db, ai_api_key="xxx")
+    news_service = NewsService(db_session=db, ai_api_key="xxx")  # 建立實例
+
+    # 如果資料庫沒新聞，先抓一次
     if db.query(NewsArticle).count() == 0:
-        # should change into simple factory pattern
-        NewsService.get_new()
-    db.close()
+        news_service.get_new()  # 用實例方法呼叫
+
+    # 排程每 100 分鐘抓新聞
     bgs.add_job(news_service.get_new, "interval", minutes=100)
     bgs.start()
-
+    db.close()
 
 @app.on_event("shutdown")
 def shutdown_scheduler():
     bgs.shutdown()
-
-class Loginverification:
-
-    def __init__(self, db_session: Session, ai_api_key: str):
-        """
-        db_session: SQLAlchemy Session
-        ai_api_key: OpenAI API Key
-        """
-        self.db = db_session
-        self.ai_key = ai_api_key
-        self.ai_client = OpenAI(api_key=self.ai_key)
-
-    pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-    oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/users/login")
-
-
-    def session_opener(self):
-        session = Session(bind=engine)
-        try:
-            yield session
-        finally:
-            session.close()
-
-
-    def verify(self,p1, p2):
-        return self.pwd_context.verify(p1, p2)
-
-
-    def check_user_password_is_correct(self,db, name, pwd):
-        user = db.query(User).filter(User.username == name).first()
-        if not self.verify(pwd, user.hashed_password):
-            return False
-        return user
-
-
-    def authenticate_user_token(
-        self,
-        token = Depends(oauth2_scheme),
-        db = Depends(session_opener)
-    ):
-        payload = jwt.decode(token, '1892dhianiandowqd0n', algorithms=["HS256"])
-        return db.query(User).filter(User.username == payload.get("sub")).first()
-
-
-    def create_access_token(self,data, expires_delta=None):
-        """create access token"""
-        to_encode = data.copy()
-        if expires_delta:
-            expire = datetime.utcnow() + expires_delta
-        else:
-            expire = datetime.utcnow() + timedelta(minutes=15)
-        to_encode.update({"exp": expire})
-        print(to_encode)
-        encoded_jwt = jwt.encode(to_encode, '1892dhianiandowqd0n', algorithm="HS256")
-        return encoded_jwt
-
-
-@app.post("/api/v1/users/login")
-async def login_for_access_token(
-        form_data: OAuth2PasswordRequestForm = Depends(), 
-        db: Session = Depends(Loginverification.session_opener)
-):
-    """login"""
-    u = Loginverification.check_user_password_is_correct(db, form_data.username, form_data.password)
-    access_token = Loginverification.create_access_token(
-        data={"sub": str(u.username)}, 
-        expires_delta=timedelta(minutes=30)
-    )
-    return {"access_token": access_token, "token_type": "bearer"}
-
-class UserAuthSchema(BaseModel):
-    username: str
-    password: str
-@app.post("/api/v1/users/register")
-def create_user(user: UserAuthSchema, db: Session = Depends(Loginverification.session_opener)):
-    """create user"""
-    hashed_password = Loginverification.pwd_context.hash(user.password)
-    db_user = User(username=user.username, hashed_password=hashed_password)
-    db.add(db_user)
-    db.commit()
-    db.refresh(db_user)
-    return db_user
-
-
-@app.get("/api/v1/users/me")
-def read_users_me(user=Depends(Loginverification.authenticate_user_token)):
-    return {"username": user.username}
-
-
-_id_counter = itertools.count(start=1000000)
-
-
-def get_article_upvote_details(article_id, uid, db):
-    cnt = (
-        db.query(user_news_association_table)
-        .filter_by(news_articles_id=article_id)
-        .count()
-    )
-    voted = False
-    if uid:
-        voted = (
-                db.query(user_news_association_table)
-                .filter_by(news_articles_id=article_id, user_id=uid)
-                .first()
-                is not None
-        )
-    return cnt, voted
-
-
-@app.get("/api/v1/news/news")
-def read_news(db=Depends(Loginverification.session_opener)):
-    """
-    read new
-
-    :param db:
-    :return:
-    """
-    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
-    result = []
-    for n in news:
-        upvotes, upvoted = get_article_upvote_details(n.id, None, db)
-        result.append(
-            {**n.__dict__, "upvotes": upvotes, "is_upvoted": upvoted}
-        )
-    return result
-
-
-@app.get("/api/v1/news/user_news")
-def read_user_news(
-        db=Depends(Loginverification.session_opener),
-        user=Depends(Loginverification.authenticate_user_token)
-):
-    """
-    read user new
-
-    :param db:
-    :param u:
-    :return:
-    """
-    news = db.query(NewsArticle).order_by(NewsArticle.time.desc()).all()
-    result = []
-    for article in news:
-        upvotes, upvoted = get_article_upvote_details(article.id, user.id, db)
-        result.append(
-            {
-                **article.__dict__,
-                "upvotes": upvotes,
-                "is_upvoted": upvoted,
-            }
-        )
-    return result
-
-class PromptRequest(BaseModel):
-    prompt: str
-
-@app.post("/api/v1/news/search_news")
-async def search_news(request: PromptRequest):
-    prompt = request.prompt
-    news_list = []
-    messages = [
-        {
-            "role": "system",
-            "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
-        },
-        {
-            "role": "user", 
-            "content": f"{prompt}"
-        },
-    ]
-
-    completion = OpenAI(api_key="xxx").chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=messages,
-    )
-    keywords = completion.choices[0].message.content
-    # should change into simple factory pattern
-    news_items = NewsService.get_new_info(keywords, is_initial=False)
-    for news in news_items:
-        try:
-            response = requests.get(news["titleLink"])
-            soup = BeautifulSoup(response.text, "html.parser")
-            # 標題
-            title = soup.find("h1", class_="article-content__title").text
-            time = soup.find("time", class_="article-content__time").text
-            # 定位到包含文章内容的 <section>
-            content_section = soup.find("section", class_="article-content__editor")
-
-            paragraphs = [
-                p.text
-                for p in content_section.find_all("p")
-                if p.text.strip() != "" and "▪" not in p.text
-            ]
-            detailed_news = {
-                "url": news["titleLink"],
-                "title": title,
-                "time": time,
-                "content": paragraphs,
-            }
-            detailed_news["content"] = " ".join(detailed_news["content"])
-            detailed_news["id"] = next(_id_counter)
-            news_list.append(detailed_news)
-        except Exception as e:
-            print(e)
-    return sorted(news_list, key=lambda x: x["time"], reverse=True)
-
-class NewsSumaryRequestSchema(BaseModel):
-    content: str
-
-@app.post("/api/v1/news/news_summary")
-async def news_summary(
-        payload: NewsSumaryRequestSchema, u=Depends(Loginverification.authenticate_user_token)
-):
-    response = {}
-    messages = [
-        {
-            "role": "system",
-            "content": "你是一個新聞摘要生成機器人，請統整新聞中提及的影響及主要原因 (影響、原因各50個字，請以json格式回答 {'影響': '...', '原因': '...'})",
-        },
-        {
-            "role": "user", 
-            "content": f"{payload.content}"
-        },
-    ]
-
-    completion = OpenAI(api_key="xxx").chat.completions.create(
-        model="gpt-3.5-turbo",
-        messages=messages,
-    )
-    result = completion.choices[0].message.content
-    if result:
-        result = json.loads(result)
-        response["summary"] = result["影響"]
-        response["reason"] = result["原因"]
-    return response
-
-
-@app.post("/api/v1/news/{id}/upvote")
-def upvote_article(
-        id,
-        db=Depends(Loginverification.session_opener),
-        u=Depends(Loginverification.authenticate_user_token),
-):
-    message = toggle_upvote(id, u.id, db)
-    return {"message": message}
-
-
-def toggle_upvote(n_id, u_id, db):
-    existing_upvote = db.execute(
-        select(user_news_association_table).where(
-            user_news_association_table.c.news_articles_id == n_id,
-            user_news_association_table.c.user_id == u_id,
-        )
-    ).scalar()
-
-    if existing_upvote:
-        delete_stmt = delete(user_news_association_table).where(
-            user_news_association_table.c.news_articles_id == n_id,
-            user_news_association_table.c.user_id == u_id,
-        )
-        db.execute(delete_stmt)
-        db.commit()
-        return "Upvote removed"
-    else:
-        insert_stmt = insert(user_news_association_table).values(
-            news_articles_id=n_id,
-            user_id=u_id
-        )
-        db.execute(insert_stmt)
-        db.commit()
-        return "Article upvoted"
-
-
-def news_exists(id2, db: Session):
-    return db.query(NewsArticle).filter_by(id=id2).first() is not None
-
-
-@app.get("/api/v1/prices/necessities-price")
-def get_necessities_prices(
-        category=Query(None), commodity=Query(None)
-):
-    return requests.get(
-        "https://opendata.ey.gov.tw/api/ConsumerProtection/NecessitiesPrice",
-        params={"CategoryName": category, "Name": commodity},
-    ).json()
