+++ backend/post/search.py
@@ -1,0 +1,61 @@
+from ..database import app,PromptRequest
+from ..NewsService import NewsService
+
+from fastapi import APIRouter
+from openai import OpenAI
+import requests
+from bs4 import BeautifulSoup
+import itertools
+
+routes=APIRouter()
+@app.post("/api/v1/news/search_news")
+async def search_news(request: PromptRequest):
+    prompt = request.prompt
+    news_list = []
+    messages = [
+        {
+            "role": "system",
+            "content": "你是一個關鍵字提取機器人，用戶將會輸入一段文字，表示其希望看見的新聞內容，請提取出用戶希望看見的關鍵字，請截取最重要的關鍵字即可，避免出現「新聞」、「資訊」等混淆搜尋引擎的字詞。(僅須回答關鍵字，若有多個關鍵字，請以空格分隔)",
+        },
+        {
+            "role": "user", 
+            "content": f"{prompt}"
+        },
+    ]
+
+    completion = OpenAI(api_key="xxx").chat.completions.create(
+        model="gpt-3.5-turbo",
+        messages=messages,
+    )
+    keywords = completion.choices[0].message.content
+    # should change into simple factory pattern
+    news_items = NewsService.get_new_info(keywords, is_initial=False)
+    for news in news_items:
+        try:
+            response = requests.get(news["titleLink"])
+            soup = BeautifulSoup(response.text, "html.parser")
+            # 標題
+            title = soup.find("h1", class_="article-content__title").text
+            time = soup.find("time", class_="article-content__time").text
+            # 定位到包含文章内容的 <section>
+            content_section = soup.find("section", class_="article-content__editor")
+            
+            _id_counter = itertools.count(start=1000000)
+
+            paragraphs = [
+                p.text
+                for p in content_section.find_all("p")
+                if p.text.strip() != "" and "▪" not in p.text
+            ]
+            detailed_news = {
+                "url": news["titleLink"],
+                "title": title,
+                "time": time,
+                "content": paragraphs,
+            }
+            detailed_news["content"] = " ".join(detailed_news["content"])
+            detailed_news["id"] = next(_id_counter)
+            news_list.append(detailed_news)
+        except Exception as e:
+            print(e)
+    return sorted(news_list, key=lambda x: x["time"], reverse=True)
